+++ Sources/Neuron/NeuroTransmitter.swift
@@ -7,28 +7,26 @@
 //
 
 import Foundation
-import UIKit
-
 
 /// Object used to create a link between Neruon objects to act as inputs
 public class NeuroTransmitter {
   public weak var neuron: Neuron?
-  public var weight: CGFloat
-  public var input: CGFloat = 0
+  public var weight: Float
+  public var input: Float = 0
   
   //as neuron connection
   public init(neuron: Neuron) {
-    self.weight = CGFloat.random(in: 0...1)
+    self.weight = Float.random(in: 0...1)
     self.neuron = neuron
   }
   
   //as standard input
-  public init(input: CGFloat) {
-    self.weight = CGFloat.random(in: 0...1)
+  public init(input: Float) {
+    self.weight = Float.random(in: 0...1)
     self.input = input
   }
   
-  public func get() -> CGFloat {
+  public func get() -> Float {
     return self.neuron?.get() ?? self.input
   }
 }
